requirements.txt
aiogram>=3.0.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
pillow>=10.0.0,<11.0.0
pdf2image>=1.17.0,<2.0.0